pyproject.toml
[project]
name = "eg4_solar_client"
version = "0.1.0"
description = "A Python client for the EG4 Solar API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
